requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0

# LLM providers
anthropic>=0.18.0
openai>=1.12.0
instructor>=0.6.0

# RSS and web scraping
feedparser>=6.0.10
requests>=2.31.0
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0

# Output
pandas>=2.0.0
gspread>=5.12.0
google-auth>=2.25.0

# Utilities
tenacity>=8.2.3
tqdm>=4.66.0

# Development
pytest>=7.4.0
pytest-mock>=3.12.0
black>=23.0.0
